Avoid showing the welcome dialog when using file -> open local file
Update the DataSourceDialog to hide itself when the active view is "file". This avoids showing the dialog when a user wants to select a file using the new app menu.

A quirk of the way we are leveraging the dialog for workspace actions is that if the user cancels the file picker they will be brought back to the welcome dialog.